+++ qrsignup/src/screens/monitor/MonitorContainer.js
@@ -1,4 +1,4 @@
-import MonitorPresenter from "./MpnitorPresenter";
+import MonitorPresenter from "./MonitorPresenter";
 
 import { handleDay } from "../../utils/Time";
 
@@ -114,13 +114,6 @@ const MonitorContainer = ({ sessionCheck }) => {
     }
   }); // 세션 유지 여부 확인, 접근 제한
 
-  const onClickLogout = () => {
-    // sessionStorage 에 user_id 로 저장되어있는 아이템을 삭제한다.
-    sessionStorage.removeItem("user_id");
-    // App 으로 이동(새로고침)
-    document.location.href = "/waiting";
-  }; // 로그아웃 기능
-
   const onClickModal2 = () => {
     usingModalChange(!usingModal);
   };
@@ -161,7 +154,6 @@ const MonitorContainer = ({ sessionCheck }) => {
       {...{ setShopNow }}
       {...{ onClickAccept }}
       {...{ adminAbout }}
-      {...{ onClickLogout }}
       {...{ sampling }}
       {...{ extendModal }}
       {...{ objectList }}
