+++ qrsignup/src/screens/monitor/MonitorPresenter.js
@@ -1,164 +1,22 @@
-import "./Monitor.css";
-import logo from "../img/kt.png";
-import icon from "../img/icon.ico";
+import logo from "../../img/kt.png";
+import icon from "../../img/icon.ico";
 
-import { handleDay } from "../utils/Time";
+import logout from "../../utils/Logout";
 
-import { useNavigate } from "react-router-dom";
-import { useState, useEffect } from "react";
-
-import {
-  getDatabase,
-  child,
-  get,
-  ref,
-  update,
-  onValue,
-} from "firebase/database";
-import { database } from "../firebase";
-
-function Monitor(props) {
-  const navigate = useNavigate();
-  const dbRef = ref(database);
-  const db = getDatabase();
-
-  const [objectList, setObjectList] = useState([]); // 객체 저장
-  let [adminAbout, setAdminAbout] = useState([]);
-  let [shopAbout, setShopAbout] = useState([]);
-  let [userId, setUserId] = useState("");
-  let [shopNow, setShopNow] = useState([]);
-
-  let [sampling, setSampling] = useState(0);
-  let [sampling1, setSampling1] = useState(0);
-  let [sampling2, setSampling2] = useState(0);
-
-  let [readTrigger, setReadTrigger] = useState(false);
-
-  let [extendModal, extendModalChange] = useState(false);
-  let [usingModal, usingModalChange] = useState(false);
-  let [notExtendedModal, notExtendedModalChange] = useState(false);
-  // 클릭시 모달 오픈/클로즈를 위한 state
-
-  useEffect(() => {
-    const resp = ref(db, "admin/" + `${sessionStorage.getItem("user_id")}`);
-
-    onValue(resp, (snapshot) => {
-      if (snapshot.exists()) {
-        setObjectList(Object.entries(snapshot.val().shop_list));
-        setAdminAbout(snapshot.val());
-      } else {
-      }
-    });
-  }, [readTrigger]); // 비동기로 처음 컴포넌트 렌더링 시에만 실행되는 hook, admin 정보 조회
-
-  useEffect(() => {
-    countFiltering();
-  }, [objectList]);
-
-  useEffect(() => {
-    if (readTrigger === true) {
-      get(child(dbRef, "shop/" + userId))
-        .then((snapshot) => {
-          if (snapshot.exists()) {
-            setShopAbout(snapshot.val());
-          } else {
-            console.log("No data available");
-          }
-        })
-        .catch((error) => {
-          console.error(error);
-        });
-    }
-  }, [readTrigger]);
-
-  useEffect(() => {
-    const updates = {};
-    if (readTrigger === true) {
-      if (shopNow.isWorking === false && shopNow.status === "onProgress") {
-        updates[
-          "admin/" +
-            `${sessionStorage.getItem("user_id")}` +
-            "/shop_list/" +
-            userId +
-            "/isWorking"
-        ] = true;
-
-        updates[
-          "admin/" +
-            `${sessionStorage.getItem("user_id")}` +
-            "/shop_list/" +
-            userId +
-            "/status"
-        ] = "isPermitted";
-
-        updates["shop/" + userId + "/expiryDate"] = handleDay(365);
-        update(ref(db), updates);
-        setReadTrigger(false);
-      } else {
-        updates[
-          "admin/" +
-            `${sessionStorage.getItem("user_id")}` +
-            "/shop_list/" +
-            userId +
-            "/status"
-        ] = "isPermitted";
-
-        updates["shop/" + userId + "/expiryDate"] = handleDay(365);
-        update(ref(db), updates);
-        setReadTrigger(false);
-      }
-    }
-  }, [shopAbout, readTrigger]); // 승인버튼 클릭시 실행
-
-  useEffect(() => {
-    if (props.isLogin === false) {
-      navigate("/");
-    }
-  }); // 세션 유지 여부 확인, 접근 제한
-
-  const onClickLogout = () => {
-    // sessionStorage 에 user_id 로 저장되어있는 아이템을 삭제한다.
-    sessionStorage.removeItem("user_id");
-    // App 으로 이동(새로고침)
-    document.location.href = "/waiting";
-  }; // 로그아웃 기능
-
-  const onClickModal2 = () => {
-    usingModalChange(!usingModal);
-  };
-
-  const onClickModal3 = () => {
-    notExtendedModalChange(!notExtendedModal);
-  };
-
-  const countArray = [0, 0, 0];
-
-  const countFiltering = () => {
-    objectList.forEach((el) => {
-      if (el[1].status === "onProgress") {
-        countArray[0]++;
-        if (el[1].isWorking === true) {
-          countArray[1]++;
-        } else if (el[1].isWorking === false) {
-          countArray[2]++;
-        }
-      } else if (el[1].isWorking === true) {
-        countArray[1]++;
-      } else if (el[1].isWorking === false) {
-        countArray[2]++;
-      }
-    });
-    setSampling(countArray[0]);
-    setSampling1(countArray[1]);
-    setSampling2(countArray[2]);
-  };
-
-  function onClickAccept(shop_id) {
-    const id = shop_id;
-    setUserId(id);
-    setReadTrigger(true);
-  }
-
+const MonitorPresenter = ({
+  setShopNow,
+  onClickAccept,
+  adminAbout,
+  sampling,
+  extendModal,
+  objectList,
+  onClickModal2,
+  sampling1,
+  usingModal,
+  sampling2,
+  notExtendedModal,
+  onClickModal3,
+}) => {
   const ExtendComp = (props) => {
     return (
       <div className="ReqBottom">
@@ -254,7 +112,7 @@ function Monitor(props) {
           </div>
           <p className="AboutAdmin">{sessionStorage.getItem("user_id")}</p>
           <p className="AboutAdmin2">{adminAbout.email}</p>
-          <button onClick={onClickLogout}>로그아웃</button>
+          <button onClick={() => logout("/")}>로그아웃</button>
         </div>
         <div className="Notification">
           <div className="NotiBox">
@@ -315,6 +173,6 @@ function Monitor(props) {
       </div>
     </div>
   );
-}
+};
 
-export default Monitor;
+export default MonitorPresenter;
